+++ mcp_tools/binary_analysis_and_reverse_engineering.py
@@ -1,0 +1,212 @@
+# mcp_tools/binary_analysis_and_reverse_engineering.py
+
+from typing import Dict, Any
+
+def register_binary_analysis_and_reverse_engineering_tools(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def gdb_analyze(binary: str, commands: str = "", script_file: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute GDB for binary analysis and debugging with enhanced logging.
+
+        Args:
+            binary: Path to the binary file
+            commands: GDB commands to execute
+            script_file: Path to GDB script file
+            additional_args: Additional GDB arguments
+
+        Returns:
+            Binary analysis results
+        """
+        data = {
+            "binary": binary,
+            "commands": commands,
+            "script_file": script_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting GDB analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/gdb", data)
+        if result.get("success"):
+            logger.info(f"✅ GDB analysis completed for {binary}")
+        else:
+            logger.error(f"❌ GDB analysis failed for {binary}")
+        return result
+
+    @mcp.tool()
+    def radare2_analyze(binary: str, commands: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Radare2 for binary analysis and reverse engineering with enhanced logging.
+
+        Args:
+            binary: Path to the binary file
+            commands: Radare2 commands to execute
+            additional_args: Additional Radare2 arguments
+
+        Returns:
+            Binary analysis results
+        """
+        data = {
+            "binary": binary,
+            "commands": commands,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Radare2 analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/radare2", data)
+        if result.get("success"):
+            logger.info(f"✅ Radare2 analysis completed for {binary}")
+        else:
+            logger.error(f"❌ Radare2 analysis failed for {binary}")
+        return result
+
+    @mcp.tool()
+    def binwalk_analyze(file_path: str, extract: bool = False, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Binwalk for firmware and file analysis with enhanced logging.
+
+        Args:
+            file_path: Path to the file to analyze
+            extract: Whether to extract discovered files
+            additional_args: Additional Binwalk arguments
+
+        Returns:
+            Firmware analysis results
+        """
+        data = {
+            "file_path": file_path,
+            "extract": extract,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Binwalk analysis: {file_path}")
+        result = hexstrike_client.safe_post("api/tools/binwalk", data)
+        if result.get("success"):
+            logger.info(f"✅ Binwalk analysis completed for {file_path}")
+        else:
+            logger.error(f"❌ Binwalk analysis failed for {file_path}")
+        return result
+
+    @mcp.tool()
+    def ropgadget_search(binary: str, gadget_type: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Search for ROP gadgets in a binary using ROPgadget with enhanced logging.
+
+        Args:
+            binary: Path to the binary file
+            gadget_type: Type of gadgets to search for
+            additional_args: Additional ROPgadget arguments
+
+        Returns:
+            ROP gadget search results
+        """
+        data = {
+            "binary": binary,
+            "gadget_type": gadget_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting ROPgadget search: {binary}")
+        result = hexstrike_client.safe_post("api/tools/ropgadget", data)
+        if result.get("success"):
+            logger.info(f"✅ ROPgadget search completed for {binary}")
+        else:
+            logger.error(f"❌ ROPgadget search failed for {binary}")
+        return result
+
+    @mcp.tool()
+    def checksec_analyze(binary: str) -> Dict[str, Any]:
+        """
+        Check security features of a binary with enhanced logging.
+
+        Args:
+            binary: Path to the binary file
+
+        Returns:
+            Security features analysis results
+        """
+        data = {
+            "binary": binary
+        }
+        logger.info(f"🔧 Starting Checksec analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/checksec", data)
+        if result.get("success"):
+            logger.info(f"✅ Checksec analysis completed for {binary}")
+        else:
+            logger.error(f"❌ Checksec analysis failed for {binary}")
+        return result
+
+    @mcp.tool()
+    def xxd_hexdump(file_path: str, offset: str = "0", length: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Create a hex dump of a file using xxd with enhanced logging.
+
+        Args:
+            file_path: Path to the file
+            offset: Offset to start reading from
+            length: Number of bytes to read
+            additional_args: Additional xxd arguments
+
+        Returns:
+            Hex dump results
+        """
+        data = {
+            "file_path": file_path,
+            "offset": offset,
+            "length": length,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting XXD hex dump: {file_path}")
+        result = hexstrike_client.safe_post("api/tools/xxd", data)
+        if result.get("success"):
+            logger.info(f"✅ XXD hex dump completed for {file_path}")
+        else:
+            logger.error(f"❌ XXD hex dump failed for {file_path}")
+        return result
+
+    @mcp.tool()
+    def strings_extract(file_path: str, min_len: int = 4, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Extract strings from a binary file with enhanced logging.
+
+        Args:
+            file_path: Path to the file
+            min_len: Minimum string length
+            additional_args: Additional strings arguments
+
+        Returns:
+            String extraction results
+        """
+        data = {
+            "file_path": file_path,
+            "min_len": min_len,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Strings extraction: {file_path}")
+        result = hexstrike_client.safe_post("api/tools/strings", data)
+        if result.get("success"):
+            logger.info(f"✅ Strings extraction completed for {file_path}")
+        else:
+            logger.error(f"❌ Strings extraction failed for {file_path}")
+        return result
+
+    @mcp.tool()
+    def objdump_analyze(binary: str, disassemble: bool = True, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Analyze a binary using objdump with enhanced logging.
+
+        Args:
+            binary: Path to the binary file
+            disassemble: Whether to disassemble the binary
+            additional_args: Additional objdump arguments
+
+        Returns:
+            Binary analysis results
+        """
+        data = {
+            "binary": binary,
+            "disassemble": disassemble,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Objdump analysis: {binary}")
+        result = hexstrike_client.safe_post("api/tools/objdump", data)
+        if result.get("success"):
+            logger.info(f"✅ Objdump analysis completed for {binary}")
+        else:
+            logger.error(f"❌ Objdump analysis failed for {binary}")
+        return result
